# SPEED: Speed up convert_np_to_gdx_svs