--- a/gdxpds/special.py
+++ b/gdxpds/special.py
@@ -130,30 +130,19 @@ def convert_np_to_gdx_svs(df, num_dims):
         their GDX equivalents
     """
 
-    # converts a single value; NANs are assumed already handled
-    def convert_approx_eps(value):
-        # eps values are not always caught by ==, use is_np_eps which applies
-        # a tolerance
-        if is_np_eps(value):
-            return SPECIAL_VALUES[4]
-        return value
-
-    # get a clean copy of df
-    try:
-        tmp = copy.deepcopy(df)
-    except:
-        logger.warning("Unable to deepcopy:\n{}".format(df))
-        tmp = copy.copy(df)
+    # get a clean copy of the values
+    values = df.iloc[:, num_dims:].copy()
 
     # fillna and apply map to value columns, then merge with dimensional columns
     try:
-        values = tmp.iloc[:, num_dims:].replace(NP_TO_GDX_SVS, value=None).applymap(convert_approx_eps)
-        tmp = (tmp.iloc[:, :num_dims]).merge(values, left_index=True, right_index=True)
+        values = values.replace(NP_TO_GDX_SVS, value=None)
+        values[(0 < values) & (values < 2 * NUMPY_SPECIAL_VALUES[-1])] = SPECIAL_VALUES[4]
+        df2 = (df.iloc[:, :num_dims]).merge(values, left_index=True, right_index=True)
     except:
         logger.error("Unable to convert numpy special values to GDX special values." + \
                      "num_dims: {}, dataframe:\n{}".format(num_dims, df))
         raise
-    return tmp
+    return df2
 
 
 def pd_isnan(val):