--- a/gdxpds/test/test_specials.py
+++ b/gdxpds/test/test_specials.py
@@ -167,3 +167,11 @@ def test_numpy_eps():
     assert(not gdxpds.special.is_np_eps(2.0 * np.finfo(float).eps))
 
 
+def test_convert_np_to_gdx_svs():
+    test_df = pds.DataFrame([["a", np.finfo(float).eps],
+                             ["b", 0.0],
+                             ["c", 2.0 * np.finfo(float).eps]],
+                            columns=["A", "Value"])
+    result_df = gdxpds.special.convert_np_to_gdx_svs(test_df, num_dims=1)
+    expected_df = pds.Series([gdxpds.special.SPECIAL_VALUES[4], 0.0, 2.0 * np.finfo(float).eps])
+    assert result_df["Value"].equals(expected_df)