@jebob 2023-05-05
Alas, I no longer have a working GAMS install and can't test this. The core technique of vectorising the replace and the EPS check should work if you wanted to fix or to re-implement it.